+++ .gitignore
@@ -8,4 +8,8 @@ temp/
 
 .vscode/
 
+bin/
+obj/
+TestResults/
+
 .fake
