+++ mutemaanpa-cs-test/CharacterTest.cs
@@ -1,0 +1,13 @@
+namespace mutemaanpa_cs_test;
+
+using Mutemaanpa;
+
+[TestClass]
+public class CharacterTest
+{
+    [TestMethod]
+    public void TestMethod1()
+    {
+        CharacterManager characterManager;
+    }
+}
