+++ mutemaanpa-cs.sln
@@ -1,12 +1,14 @@
 ﻿Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mutemaanpa-cs", "mutemaanpa-cs.csproj", "{C07105E5-C0DE-4E30-8125-4C16FD3A1FA9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mutemaanpa-cs", "mutemaanpa-cs\mutemaanpa-cs.csproj", "{C07105E5-C0DE-4E30-8125-4C16FD3A1FA9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mutemaanpa-cs-test", "mutemaanpa-cs-test\mutemaanpa-cs-test.csproj", "{99FFE4AD-7448-4759-9557-F8F65871ACD7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-	Debug|Any CPU = Debug|Any CPU
-	ExportDebug|Any CPU = ExportDebug|Any CPU
-	ExportRelease|Any CPU = ExportRelease|Any CPU
+		Debug|Any CPU = Debug|Any CPU
+		ExportDebug|Any CPU = ExportDebug|Any CPU
+		ExportRelease|Any CPU = ExportRelease|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{C07105E5-C0DE-4E30-8125-4C16FD3A1FA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -15,5 +17,11 @@ Global
 		{C07105E5-C0DE-4E30-8125-4C16FD3A1FA9}.ExportDebug|Any CPU.Build.0 = ExportDebug|Any CPU
 		{C07105E5-C0DE-4E30-8125-4C16FD3A1FA9}.ExportRelease|Any CPU.ActiveCfg = ExportRelease|Any CPU
 		{C07105E5-C0DE-4E30-8125-4C16FD3A1FA9}.ExportRelease|Any CPU.Build.0 = ExportRelease|Any CPU
+		{99FFE4AD-7448-4759-9557-F8F65871ACD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99FFE4AD-7448-4759-9557-F8F65871ACD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{99FFE4AD-7448-4759-9557-F8F65871ACD7}.ExportDebug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99FFE4AD-7448-4759-9557-F8F65871ACD7}.ExportDebug|Any CPU.Build.0 = Debug|Any CPU
+		{99FFE4AD-7448-4759-9557-F8F65871ACD7}.ExportRelease|Any CPU.ActiveCfg = Debug|Any CPU
+		{99FFE4AD-7448-4759-9557-F8F65871ACD7}.ExportRelease|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 EndGlobal
