--- script/character/CharacterManager.cs
@@ -1,9 +1,0 @@
-namespace Mutemaanpa
-{
-    /// <summary>
-    /// Holds all in-game characters data, and persists them.
-    /// </summary>
-    public class CharacterManager(Database database) {
-        
-    }
-}
